parse: Give node members more expressive names
Members of nodes have names such as "logical-OR-pkgex" that describe
what the members are, but not what their meaning is.

This commit modifies the parser so that it gives node members names
that describe what the members' meaning is. For example, the members
of logical-OR-pkgex nodes are renamed to left_child and right_child
to represent their order of evaluation.