Use tabs for indentation, spaces for alignment
The indentation and alignment of some of the code is incorrect, using
spaces for indentation, and tabs for alignment. This causes the code
to be displayed incorrectly when viewed with a different tab width.

This commit modifies all source files so that tabs and spaces are
used consistently for indentation and alignment, respectively.